tpm: an initial abstractiong of hw interface
Tboot, from which the code came, is only concerned for x86 systems. Typically
on x86 systems, the TPM is accessed using the MMIO hardware interface for
which there are two ABIs, FIFO and CRB.

This commit looks to refactor the hardware interface logic to enable the
ability to support other hardware interfaces, such as the Mobile TPM interface
typically used on Arm.
