+++ xen/drivers/security/tpm/Makefile
@@ -1,3 +1,4 @@
 obj-y += tpm.o
+obj-y += tpm_mmio.o
 obj-y += tpm_12.o
 obj-y += tpm_20.o
